CSIT 998 Casptone Project (Power BI with Automated Data-driven Analysis)/source/dashboard-powerbi/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
joblib>=1.2.0
PyMySQL>=1.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-multipart>=0.0.9
requests>=2.31.0
msal>=1.27.0
python-dotenv>=1.0.1
